+++ src/module/applications/api/document-sheet-mixin.mjs
@@ -44,8 +44,9 @@ export default base => {
     /**
      * The mode the sheet is currently in.
      * @type {DSDocumentSheet.MODES}
+     * @protected
      */
-    _mode;
+    _mode = DSDocumentSheet.MODES.PLAY;
 
     /* -------------------------------------------------- */
 
@@ -73,6 +74,8 @@ export default base => {
     _configureRenderOptions(options) {
       super._configureRenderOptions(options);
       if (options.mode && this.isEditable) this._mode = options.mode;
+      // New sheets should always start in edit mode
+      else if (options.renderContext === `create${this.document.documentName}`) this._mode = DSDocumentSheet.MODES.EDIT;
     }
 
     /* -------------------------------------------------- */