+++ CHANGELOG.md
@@ -23,12 +23,31 @@
 ### Added
 
 - Equipment now has a "quantity" property. Completing a project will now add to an equipment's quantity if it already exists on the actor, based on matching `dsid` values. (#558)
+- Added Healing enricher, which lets you recover stamina or temporary stamina with `[[/heal]]`. See the wiki for more details. (#559)
+- Added Damage enricher, which lets you add damage rolls to text with `[[/damage]]`. See the wiki for more details. (#568)
+- Add Adjust Malice and Reset Malice context menu options to the player UI. (#515)
 
 ### Changed
 
-- Adjusted display of the actor sheets in play mode
+- Overhauled the Effects tab of the Item & Actor sheets. (#146)
+  - Actor sheets now feature a "Status Conditions" section which lists the conditions specific to Draw Steel.
+    - Clicking the buttons will toggle the status
+    - If you have a status, it will be highlighted in orange
+    - If the status is granted by a non-canonical condition, the button will be disabled
+  - Hitting "New Effect" in the "Temporary Effects" section will now set the duration to End of Turn, instead of 1 round.
+  - An effect's enriched description is now available by toggling the carat, like already existed for items.
+- Replaced the default token status menu with a new one that allows applying a status with one of Draw Steel's unique durations. (#261)
+- Adjusted display of the actor sheets in play mode. (#609)
   - Unused ability and equipment sections are no longer displayed in play mode.
   - Add Item buttons only display in edit mode.
+- New actors and items now start in edit mode. Items no longer default to edit mode if opened while editable.
+- Show malice value in the player UI on combat creation instead of only on beginning combat. (#498)
+- Moved ability bonus calculations to after active effect application. (#613)
+
+### Fixed
+
+- Fixed various color issues on sheets that had their theme set individually. (#599)
+- Fixed malice generation applying the first round bonus twice.
 
 ## 0.7.2